develop -> main

+++ src/arch/epoll.h
@@ -1,28 +1,29 @@
 #ifndef NOSTR_INTERNAL_SEND_H_
 #define NOSTR_INTERNAL_SEND_H_
 
+#include "../util/types.h"
 #include "./linux/epoll.h"
 #include "./linux/x86_64/epoll.h"
 
-static inline long internal_epoll_ctl(
-    const int            epoll_fd,
-    const int            op,
-    const int            fd,
+static inline int32_t internal_epoll_ctl(
+    const int32_t        epoll_fd,
+    const int32_t        op,
+    const int32_t        fd,
     PWebSocketEpollEvent event)
 {
     return linux_x8664_epoll_ctl(epoll_fd, op, fd, event);
 }
 
-static inline long internal_epoll_create1(const int flags)
+static inline int32_t internal_epoll_create1(const int32_t flags)
 {
     return linux_x8664_epoll_create1(flags);
 }
 
-static inline long internal_epoll_wait(
-    const int            epfd,
+static inline int32_t internal_epoll_wait(
+    const int32_t        epfd,
     PWebSocketEpollEvent events,
-    const int            maxevents,
-    const int            timeout)
+    const int32_t        maxevents,
+    const int32_t        timeout)
 {
     return linux_x8664_epoll_wait(epfd, events, maxevents, timeout);
 }
