+++ src/arch/linux/x86_64/accept.h
@@ -5,16 +5,16 @@
 #include "../sockaddr.h"
 #include "./asm.h"
 
-static inline int linux_x8664_accept4(const int sock_fd, const struct sockaddr* addr, const socklen_t* addrlen, const int flags)
+static inline int32_t linux_x8664_accept4(const int32_t sock_fd, const struct sockaddr* addr, const socklen_t* addrlen, const int32_t flags)
 {
-    long ret = linux_x8664_asm_syscall4(
+    int32_t ret = linux_x8664_asm_syscall4(
         __NR_accept4,
         sock_fd,
         addr,
         addrlen,
         flags);
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         ret   = -1;
     }
