+++ src/arch/linux/x86_64/close.h
@@ -4,13 +4,13 @@
 #include "../errno.h"
 #include "./asm.h"
 
-static inline int linux_x8664_close(const int fd)
+static inline int32_t linux_x8664_close(const int fd)
 {
-    long ret = linux_x8664_asm_syscall1(
+    int32_t ret = linux_x8664_asm_syscall1(
         __NR_close,
         fd);
 
-    if ((unsigned long)ret >= (unsigned long)-4095) {
+    if ((uint64_t)ret >= (uint64_t)-4095) {
         errno = -ret;
         ret   = -1;
     }
